requirements.txt
Django==6.0
requests==2.32.3
python-dotenv==1.0.1
PyMuPDF==1.24.13
whitenoise==6.8.2
